patient parameters refactoring

+++ src/Services/Agents.API/Agents.API.Entities/PredictRequest.cs
@@ -13,6 +13,6 @@ namespace Agents.API.Entities
 
         public string Version { get; set; }
 
-        public float[] Input { get; set; }
+        public double[] Input { get; set; }
     }
 }
